+++ libs/common/rendersettings_utils.cpp
@@ -7,6 +7,7 @@
 #include "rendersettings_utils.h"
 
 #include "constant_strings.h"
+#include "common_utils.h"
 #include "parameters_utils.h"
 
 
@@ -100,10 +101,9 @@ ArnoldAOVTypes GetArnoldTypesFromFormatToken(const TfToken &type)
     }
 }
 
-// TODO: this is duplicated code, remove it 
 // Read eventual connections to a ArnoldNodeGraph primitive, that acts as a passthrough
 static inline void UsdArnoldNodeGraphConnection(AtNode *node, const UsdPrim &prim, const UsdAttribute &attr,
-                                                const std::string &attrName, ArnoldAPIAdapter &context, TimeSettings &time)
+                                                const std::string &attrName, ArnoldAPIAdapter &context, const TimeSettings &time)
 {
     VtValue value;
     if (attr && attr.Get(&value, time.frame)) {
@@ -137,6 +137,52 @@ static inline void UsdArnoldNodeGraphConnection(AtNode *node, const UsdPrim &pri
 }
 
 
+// Read eventual connections to a ArnoldNodeGraph primitive for the aov_shader shader array connections
+static inline void UsdArnoldNodeGraphAovConnection(AtNode *options, const UsdPrim &prim, 
+    const UsdAttribute &attr, const std::string &attrBase, ArnoldAPIAdapter &context, const TimeSettings &time)
+{
+
+    VtValue value;
+    if (attr && attr.Get(&value, time.frame)) {
+        // RenderSettings have a string attribute, referencing multiple prims in the stage
+        std::string valStr = VtValueGetString(value);
+        if (!valStr.empty()) {
+            AtArray* aovShadersArray = AiNodeGetArray(options, str::aov_shaders);
+            unsigned numElements = AiArrayGetNumElements(aovShadersArray);
+            for(const auto &nodeGraphPrimName: TfStringTokenize(valStr)) {
+                SdfPath nodeGraphPrimPath(nodeGraphPrimName);
+                // We check if there is a primitive at the path of this string
+                UsdPrim nodeGraphPrim = prim.GetStage()->GetPrimAtPath(nodeGraphPrimPath);
+
+                if (nodeGraphPrim && nodeGraphPrim.GetTypeName() == _tokens->ArnoldNodeGraph) {
+                    // We can use a UsdShadeShader schema in order to read connections
+                    UsdShadeShader ngShader(nodeGraphPrim);
+                    for (unsigned aovShaderIndex=1;; aovShaderIndex++) {
+                        // the output terminal name will be aov_shader:i{1,...,n} as a contiguous array
+                        TfToken outputName(attrBase + std::string(":i") + std::to_string(aovShaderIndex));
+                        UsdShadeOutput outputAttr = ngShader.GetOutput(outputName);
+                        if (!outputAttr) {
+                            break;
+                        }
+                        SdfPathVector sourcePaths;
+                        // Check which shader is connected to this output
+                        if (outputAttr.HasConnectedSource() && outputAttr.GetRawConnectedSourcePaths(&sourcePaths)) {
+                            for(const SdfPath &aovShaderPath : sourcePaths) {
+                                SdfPath aovShaderPrimPath(aovShaderPath.GetPrimPath());
+                                UsdPrim outPrim = prim.GetStage()->GetPrimAtPath(aovShaderPrimPath);
+                                if (outPrim) {
+                                    // we connect to aov_shaders{0,...,n-1} parameters i.e. 0 indexed, offset from any previous connections
+                                    std::string optionAovShaderElement = attrBase + "[" + std::to_string(numElements++) + "]";
+                                    context.AddConnection(options, optionAovShaderElement, aovShaderPrimPath.GetText(), ArnoldAPIAdapter::CONNECTION_PTR);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
 
 // Encapsulate the logic to extract driver type and settings from a UsdProduct prim
 // The function can return nullptr if it wasn't able to find the driver
@@ -172,8 +218,8 @@ AtNode * ReadDriverFromRenderProduct(const UsdRenderProduct &renderProduct, Arno
             }
             const int paramType = AiParamGetType(paramEntry); 
             const int arrayType = AiParamGetSubType(paramEntry);
-            ReadAttribute(attr, driver, driverParamName, time, context, paramType, arrayType);
-
+            ReadAttribute(attr, driver, driverParamName, time,
+                           context, paramType, arrayType);
         }
     }
 
@@ -239,12 +285,12 @@ AtNode * DeduceDriverFromFilename(const UsdRenderProduct &renderProduct, ArnoldA
             }
             const int paramType = AiParamGetType(paramEntry); 
             const int arrayType = AiParamGetSubType(paramEntry);
-            ReadAttribute(attr, driver, driverParamName, time, context, paramType, arrayType);
+            ReadAttribute(attr, driver, driverParamName, time,
+                        context, paramType, arrayType);
         }
     }
     return driver;
 }
-
 
 
 // THIS IS NOT USED in the render delegate
@@ -358,12 +404,19 @@ void ChooseRenderSettings(UsdStageRefPtr _stage, std::string &_renderSettings, T
 }
 
 
-void ReadCameraSettings(const UsdRenderSettings &renderSettings, ArnoldAPIAdapter &context, const TimeSettings &time, AtUniverse *universe, SdfPath& camera) {
+
+AtNode* ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &context, 
+    const TimeSettings &time, AtUniverse *universe, SdfPath& cameraPath) {
+
     AtNode *options = AiUniverseGetOptions(universe);
-    auto stage = renderSettings.GetPrim().GetStage();
+    auto stage = renderSettingsPrim.GetStage();
+    UsdRenderSettings renderSettings(renderSettingsPrim);
+    if (!renderSettings)
+        return nullptr;
+
     VtValue pixelAspectRatioValue;
     if (renderSettings.GetPixelAspectRatioAttr().Get(&pixelAspectRatioValue, time.frame))
-        AiNodeSetFlt(options, str::pixel_aspect_ratio, pixelAspectRatioValue.Get<float>());
+        AiNodeSetFlt(options, str::pixel_aspect_ratio, VtValueGetFloat(pixelAspectRatioValue));
     
     GfVec2i resolution; 
     if (renderSettings.GetResolutionAttr().Get(&resolution, time.frame)) {
@@ -378,6 +431,11 @@ void ReadCameraSettings(const UsdRenderSettings &renderSettings, ArnoldAPIAdapte
         resolution[0] = AiNodeGetInt(options, str::xres);
         resolution[1] = AiNodeGetInt(options, str::yres);
     }
+    // Set default attribute values so that they match the defaults in arnold plugins, 
+    // as well as the render delegate's #1525
+    AiNodeSetInt(options, str::AA_samples, 3);
+    AiNodeSetInt(options, str::GI_diffuse_depth, 1);
+    AiNodeSetInt(options, str::GI_specular_depth, 1);
 
     // Eventual render region: in arnold it's expected to be in pixels in the range [0, resolution]
     // but in usd it's between [0, 1]
@@ -409,35 +467,16 @@ void ReadCameraSettings(const UsdRenderSettings &renderSettings, ArnoldAPIAdapte
     // instantShutter will ignore any motion blur
     VtValue instantShutterValue;
     if (renderSettings.GetInstantaneousShutterAttr().Get(&instantShutterValue, time.frame) && 
-            instantShutterValue.Get<bool>()) {
+            VtValueGetBool(instantShutterValue)) {
         AiNodeSetBool(options, str::ignore_motion_blur, true);
     }
 
     // Get the camera used for rendering, this is needed in arnold
-    UsdRelationship cameraRel = renderSettings.GetCameraRel();
-    SdfPathVector camTargets;
-    cameraRel.GetTargets(&camTargets);
-    if (!camTargets.empty()) {
-        camera = camTargets[0];
-        context.AddConnection(options, "camera", camera.GetText(), ArnoldAPIAdapter::CONNECTION_PTR);
-    }
-}
-
-void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &context, const TimeSettings &time, AtUniverse *universe, SdfPath& camera) {
-    AtNode *options = AiUniverseGetOptions(universe);
-    auto stage = renderSettingsPrim.GetStage();
-    UsdRenderSettings renderSettings(renderSettingsPrim);
-    if (!renderSettings)
-        return;
-
-    // Set default attribute values so that they match the defaults in arnold plugins, 
-    // as well as the render delegate's #1525
-    AiNodeSetInt(options, str::AA_samples, 3);
-    AiNodeSetInt(options, str::GI_diffuse_depth, 1);
-    AiNodeSetInt(options, str::GI_specular_depth, 1);
-
-    ReadCameraSettings(renderSettings, context, time, universe, camera);
-
+    UsdPrim camera = renderSettingsPrim.GetStage()->GetPrimAtPath(cameraPath);
+    // just supporting a single camera for now
+    if (camera)
+        context.AddConnection(options, "camera", camera.GetPath().GetText(), ArnoldAPIAdapter::CONNECTION_PTR);
+    
     std::vector<std::string> outputs;
     std::vector<std::string> lpes;
     std::vector<AtNode *> aovShaders;
@@ -449,7 +488,8 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
     SdfPathVector productTargets;
     productsRel.GetTargets(&productTargets);
     for (size_t i = 0; i < productTargets.size(); ++i) {
-        UsdPrim productPrim = stage->GetPrimAtPath(productTargets[i]);
+
+        UsdPrim productPrim = renderSettingsPrim.GetStage()->GetPrimAtPath(productTargets[i]);
         UsdRenderProduct renderProduct(productPrim);
         if (!renderProduct) // couldn't find the render product in the usd scene
             continue;
@@ -485,7 +525,7 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
         bool isDriverExr = AiNodeIs(driver, str::driver_exr);
         for (size_t j = 0; j < renderVarsTargets.size(); ++j) {
 
-            UsdPrim renderVarPrim = stage->GetPrimAtPath(renderVarsTargets[j]);
+            UsdPrim renderVarPrim = renderSettingsPrim.GetStage()->GetPrimAtPath(renderVarsTargets[j]);
             if (!renderVarPrim || !renderVarPrim.IsActive())
                 continue;
             UsdRenderVar renderVar(renderVarPrim);
@@ -500,8 +540,9 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
             UsdAttribute filterAttr = renderVarPrim.GetAttribute(_tokens->aovSettingFilter);
             if (filterAttr) {
                 VtValue filterValue;
-                if (filterAttr.Get(&filterValue, time.frame))
+                if (filterAttr.Get(&filterValue, time.frame)) {
                     filterType = VtValueGetString(filterValue);
+                }
             }
 
             // Create a filter node of the given type
@@ -515,14 +556,14 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
                 UsdAttribute filterWidthAttr = renderVarPrim.GetAttribute(_tokens->aovSettingWidth);
                 VtValue filterWidthValue;
                 if (filterWidthAttr && filterWidthAttr.Get(&filterWidthValue, time.frame)) {
-                    AiNodeSetFlt(filter, str::width, filterWidthValue.Get<float>());
+                    AiNodeSetFlt(filter, str::width, VtValueGetFloat(filterWidthValue));
                 }
             }
 
             // read attributes for a specific filter type, authored as "arnold:gaussian_filter:my_attr"
             std::string filterTypeAttrs = "arnold:";
             filterTypeAttrs += filterType;
-            //ReadArnoldParameters(renderVarPrim, context, filter, time, TfToken(filterTypeAttrs.c_str()));
+            ReadArnoldParameters(renderVarPrim, context, filter, time, TfToken(filterTypeAttrs.c_str()));
             filterName = AiNodeGetName(filter);
 
             TfToken dataType;
@@ -543,7 +584,7 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
             // Get the name for this AOV
             VtValue sourceNameValue;
             std::string sourceName = renderVar.GetSourceNameAttr().Get(&sourceNameValue, time.frame) ?
-                sourceNameValue.Get<std::string>() : "RGBA";
+                VtValueGetString(sourceNameValue) : "RGBA";
 
             // we want to consider "color" as referring to the beauty, just like "RGBA" (see #1311)
             if (sourceName == "color")
@@ -559,19 +600,19 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
 
             // read the parameter "driver:parameters:aov:name" that will be needed if we have merged exrs (see #816)
             if (renderVarPrim.GetAttribute(_tokens->aovSettingName).Get(&aovNameValue, time.frame)) {
-                std::string aovNameValueStr = aovNameValue.Get<std::string>();
+                std::string aovNameValueStr = VtValueGetString(aovNameValue);
                 if (!aovNameValueStr.empty()) {
                     layerName = aovNameValueStr;
                     hasLayerName = true;
                 }
             }
-
             // optional per-AOV camera
             std::string cameraName;
             VtValue cameraValue;
             if (renderVarPrim.GetAttribute(_tokens->aovSettingCamera).Get(&cameraValue, time.frame)) {
                 cameraName = VtValueGetString(cameraValue);
             }            
+            
             std::string aovName = sourceName;
             
             if (sourceType == UsdRenderTokens->lpe) {
@@ -594,8 +635,7 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
 
                 // Create a user data shader that will read the desired primvar, its type depends on the AOV type
                 std::string userDataName = renderVarPrim.GetPath().GetText() + std::string("/user_data");
-                AtNode *userData = context.CreateArnoldNode(arnoldTypes.userData.c_str(), userDataName.c_str());
-                
+                AtNode *userData = context.CreateArnoldNode(arnoldTypes.userData, userDataName.c_str());
                 // Link the user_data to the aov_write
                 AiNodeLink(userData, "aov_input", aovShader);
                 // Set the user data (primvar) to read
@@ -627,11 +667,11 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
             std::string output;
             if (!cameraName.empty())
                 output = cameraName + std::string(" ");
-
             output += aovName; // AOV name
             output += std::string(" ") + arnoldTypes.outputString; // AOV type (RGBA, VECTOR, etc..)
             output += std::string(" ") + filterName; // name of the filter for this AOV
             output += std::string(" ") + driverName; // name of the driver for this AOV
+
             // Track beauty outputs drivers
             if (aovName == "RGBA")
                 beautyDrivers.insert(driver);
@@ -710,13 +750,13 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
     ReadArnoldParameters(renderSettingsPrim, context, options, time, "arnold:global");
 
     // Read eventual connections to a node graph
-    // TODO: UsdArnoldNodeGraphConnection
-    // UsdArnoldNodeGraphConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalAtmosphere), "atmosphere", context, time);
-    // UsdArnoldNodeGraphConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalBackground), "background", context, time);
-    // UsdArnoldNodeGraphAovConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalAovs), "aov_shaders", context);
-    // for (auto driver: beautyDrivers) {
-    //     UsdArnoldNodeGraphConnection(driver, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalImager), "input", context, time);
-    // }
+    UsdArnoldNodeGraphConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalAtmosphere), "atmosphere", context, time);
+    UsdArnoldNodeGraphConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalBackground), "background", context, time);
+    UsdArnoldNodeGraphAovConnection(options, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalAovs), "aov_shaders", context, time);
+
+    for (auto driver: beautyDrivers) {
+        UsdArnoldNodeGraphConnection(driver, renderSettingsPrim, renderSettingsPrim.GetAttribute(_tokens->aovGlobalImager), "input", context, time);
+    }
 
     // Setup color manager
     AtNode* colorManager = nullptr;
@@ -779,7 +819,7 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
     if (UsdAttribute logFileAttr = renderSettingsPrim.GetAttribute(_tokens->logFile)) {
         VtValue logFileValue;
         if (logFileAttr.Get(&logFileValue, time.frame)) {
-            std::string logFile = logFileValue.Get<std::string>();
+            std::string logFile = VtValueGetString(logFileValue);
             AiMsgSetLogFileName(logFile.c_str());
         }
     }
@@ -788,12 +828,18 @@ void ReadRenderSettings(const UsdPrim &renderSettingsPrim, ArnoldAPIAdapter &con
     if (UsdAttribute logVerbosityAttr = renderSettingsPrim.GetAttribute(_tokens->logVerbosity)) {
         VtValue logVerbosityValue;
         if (logVerbosityAttr.Get(&logVerbosityValue, time.frame)) {
-            // TODO
-            // int logVerbosity = ArnoldUsdGetLogVerbosityFromFlags(logVerbosityValue.Get<int>());
-            // AiMsgSetConsoleFlags(AiNodeGetUniverse(options), logVerbosity);
-            // AiMsgSetLogFileFlags(AiNodeGetUniverse(options), logVerbosity);
+            int logVerbosity = ArnoldUsdGetLogVerbosityFromFlags(VtValueGetInt(logVerbosityValue));
+#if ARNOLD_VERSION_NUM >= 70100
+            AiMsgSetConsoleFlags(AiNodeGetUniverse(options), logVerbosity);
+            AiMsgSetLogFileFlags(AiNodeGetUniverse(options), logVerbosity);
+#else
+            AiMsgSetConsoleFlags(nullptr, logVerbosity);
+            AiMsgSetLogFileFlags(nullptr, logVerbosity);
+#endif
         }
     }
+    
+    return options;
 }
 
 PXR_NAMESPACE_CLOSE_SCOPE